Refactor: 정의되지 않은 타입 정리

+++ src/components/ProjectInfo.tsx
@@ -27,8 +27,14 @@ const ProjectInfo = () => {
   }
 
   const onDeleteAll = () => {
-    dispatch(deleteAll(Object.keys(checkItems).map(Number)))
-    unCheckAll();
+    let message = "정말 삭제하시겠습니까?";
+    let result = window.confirm(message);
+    if (result) {
+      dispatch(deleteAll(Object.keys(checkItems).map(Number)))
+      unCheckAll();
+    } else {
+      return;
+    }
   }
 
   return (
